requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
groq>=0.9.0
chromadb>=0.5.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
